changed layout & style, added logo and fixed user info missing when unregistered

+++ view/forum/listTopics.php
@@ -22,15 +22,17 @@ if (isset($topics)) {
             <?= $topic->displayHeureCreation(); ?>
 
             <?php
-            $user = $_SESSION['user'];
-            if (serialize($user) == serialize($topic->getUser())) { ?>
-                <a class='delete-btn btn btn-danger p-1'
-                    href="index.php?ctrl=forum&action=deleteTopic&cat_id=<?= $category->getId() ?>&topic_id=<?= $topic->getId(); ?>"><i
-                        class="fa-solid fa-x"></i></a>
-                <?php
-            } ?>
-        </p>
-    <?php }
+            if (isset($_SESSION['user'])) {
+                $user = $_SESSION['user'];
+                if (serialize($user) == serialize($topic->getUser())) { ?>
+                    <a class='delete-btn btn btn-danger p-1'
+                        href="index.php?ctrl=forum&action=deleteTopic&cat_id=<?= $category->getId() ?>&topic_id=<?= $topic->getId(); ?>"><i
+                            class="fa-solid fa-x"></i></a>
+                    <?php
+                } ?>
+            </p>
+        <?php }
+    }
 } else { ?>
     <p>Aucun Topic</p>
 <?php }
@@ -52,14 +54,7 @@ if (isset($_SESSION['user'])) {
 
                 <p>
                     <label class="form-label">
-                        Contenu :
-                        <textarea name="contenu" rows='5' col='33' class="form-control"></textarea>
-                    </label>
-                </p>
-
-                <p>
-                    <label class="form-label">
-                        <input class="btn btn-primary" type="submit" name="submit" value="Poster">
+                        <input class="btn btn-dark" type="submit" name="submit" value="Poster">
                     </label>
                 </p>
             </form>
